feat: replace references to camera uploads with automatic uploads in the source code
