+++ owncloudApp/src/main/java/com/owncloud/android/presentation/settings/SettingsActivity.kt
@@ -31,8 +31,8 @@ import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.core.view.isVisible
 import com.owncloud.android.R
 import com.owncloud.android.presentation.settings.advanced.SettingsAdvancedFragment
-import com.owncloud.android.presentation.settings.autouploads.SettingsPictureUploadsFragment
-import com.owncloud.android.presentation.settings.autouploads.SettingsVideoUploadsFragment
+import com.owncloud.android.presentation.settings.automaticuploads.SettingsPictureUploadsFragment
+import com.owncloud.android.presentation.settings.automaticuploads.SettingsVideoUploadsFragment
 import com.owncloud.android.presentation.settings.logging.SettingsLogsFragment
 import com.owncloud.android.presentation.settings.more.SettingsMoreFragment
 import com.owncloud.android.presentation.settings.security.SettingsSecurityFragment
