+++ owncloudDomain/src/main/java/com/owncloud/android/domain/automaticuploads/model/AutomaticUploadsConfiguration.kt
@@ -16,17 +16,11 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.owncloud.android.domain.camerauploads.usecases
+package com.owncloud.android.domain.automaticuploads.model
 
-import com.owncloud.android.domain.BaseUseCase
-import com.owncloud.android.domain.camerauploads.FolderBackupRepository
-import com.owncloud.android.domain.camerauploads.model.FolderBackUpConfiguration
-import kotlinx.coroutines.flow.Flow
-
-class GetPictureUploadsConfigurationStreamUseCase(
-    private val folderBackupRepository: FolderBackupRepository
-) : BaseUseCase<Flow<FolderBackUpConfiguration?>, Unit>() {
-
-    override fun run(params: Unit): Flow<FolderBackUpConfiguration?> =
-        folderBackupRepository.getFolderBackupConfigurationByNameAsFlow(FolderBackUpConfiguration.pictureUploadsName)
+data class AutomaticUploadsConfiguration(
+    val pictureUploadsConfiguration: FolderBackUpConfiguration?,
+    val videoUploadsConfiguration: FolderBackUpConfiguration?
+) {
+    fun areAutomaticUploadsDisabled() = pictureUploadsConfiguration == null && videoUploadsConfiguration == null
 }
