+++ owncloudDomain/src/main/java/com/owncloud/android/domain/automaticuploads/usecases/GetPictureUploadsConfigurationStreamUseCase.kt
@@ -16,16 +16,17 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.owncloud.android.domain.camerauploads.usecases
+package com.owncloud.android.domain.automaticuploads.usecases
 
 import com.owncloud.android.domain.BaseUseCase
-import com.owncloud.android.domain.camerauploads.FolderBackupRepository
-import com.owncloud.android.domain.camerauploads.model.FolderBackUpConfiguration.Companion.videoUploadsName
+import com.owncloud.android.domain.automaticuploads.FolderBackupRepository
+import com.owncloud.android.domain.automaticuploads.model.FolderBackUpConfiguration
+import kotlinx.coroutines.flow.Flow
 
-class ResetVideoUploadsUseCase(
+class GetPictureUploadsConfigurationStreamUseCase(
     private val folderBackupRepository: FolderBackupRepository
-) : BaseUseCase<Unit, Unit>() {
+) : BaseUseCase<Flow<FolderBackUpConfiguration?>, Unit>() {
 
-    override fun run(params: Unit) =
-        folderBackupRepository.resetFolderBackupConfigurationByName(videoUploadsName)
+    override fun run(params: Unit): Flow<FolderBackUpConfiguration?> =
+        folderBackupRepository.getFolderBackupConfigurationByNameAsFlow(FolderBackUpConfiguration.pictureUploadsName)
 }
