+++ owncloudDomain/src/main/java/com/owncloud/android/domain/transfers/model/UploadEnqueuedBy.kt
@@ -29,6 +29,6 @@ package com.owncloud.android.domain.transfers.model
  */
 enum class UploadEnqueuedBy {
     ENQUEUED_BY_USER,
-    ENQUEUED_AS_CAMERA_UPLOAD_PICTURE,
-    ENQUEUED_AS_CAMERA_UPLOAD_VIDEO;
+    ENQUEUED_AS_AUTOMATIC_UPLOAD_PICTURE,
+    ENQUEUED_AS_AUTOMATIC_UPLOAD_VIDEO;
 }
