+++ owncloudTestUtil/src/main/java/com/owncloud/android/testutil/OCTransfer.kt
@@ -20,7 +20,7 @@
 
 package com.owncloud.android.testutil
 
-import com.owncloud.android.domain.camerauploads.model.UploadBehavior
+import com.owncloud.android.domain.automaticuploads.model.UploadBehavior
 import com.owncloud.android.domain.transfers.model.OCTransfer
 import com.owncloud.android.domain.transfers.model.TransferStatus
 import com.owncloud.android.domain.transfers.model.UploadEnqueuedBy
